net: ionic: Export the APIs from net driver to support device commands
RDMA driver needs to establish admin queues to support admin operations.
Export the APIs to send device commands for the RDMA driver.
